feat: Create a subpackage for types that are useful elsewhere
