+++ src/conductor.ts
@@ -11,7 +11,7 @@ import { Device, DeviceOptions } from './devices/device'
 import { CasparCGDevice } from './devices/casparCG'
 import { AbstractDevice } from './devices/abstract'
 import { HttpSendDevice } from './devices/httpSend'
-import { Mappings, Mapping, DeviceType, TimelineResolvedObjectExtended } from './devices/mapping'
+import { Mappings, Mapping, DeviceType, TimelineResolvedObjectExtended } from './types/mapping'
 import { AtemDevice } from './devices/atem'
 import { EventEmitter } from 'events'
 import { LawoDevice } from './devices/lawo'
