+++ src/devices/panasonicPTZ.ts
@@ -8,21 +8,19 @@ import {
 } from './device'
 import {
 	DeviceType,
-	MappingPanasonicPtz,
-	Mappings,
-	MappingPanasonicPtzType
-} from './mapping'
-import {
-	TimelineContentTypePanasonicPtz
-} from '../enums'
+	Mappings
+} from '../types/mapping'
 import {
 	TimelineObjPanasonicPtz,
 	TimelineObjPanasonicPtzPreset,
 	TimelineObjPanasonicPtzPresetSpeed,
 	TimelineObjPanasonicPtzZoomSpeed,
-	TimelineObjPanasonicPtzZoom
-} from '../types'
-import { TimelineState } from 'superfly-timeline'
+	TimelineObjPanasonicPtzZoom,
+	TimelineContentTypePanasonicPtz,
+	MappingPanasonicPtz,
+	MappingPanasonicPtzType
+} from '../types/ptz'
+import { TimelineState, TimelineResolvedObject } from 'superfly-timeline'
 import { DoOnTime } from '../doOnTime'
 import { PanasonicPtzHttpInterface } from './panasonicPTZAPI'
 
@@ -122,26 +120,26 @@ export class PanasonicPtzDevice extends DeviceWithState<TimelineState> {
 		// convert the timeline state into something we can use
 		const ptzState: PanasonicPtzState = this._getDefaultState()
 
-		_.each(state.LLayers, (tlObject: TimelineObjPanasonicPtz, layerName: string) => {
+		_.each(state.LLayers, (tlObject: TimelineResolvedObject & TimelineObjPanasonicPtz, layerName: string) => {
 			const mapping: MappingPanasonicPtz | undefined = this.mapping[layerName] as MappingPanasonicPtz
 			if (mapping && mapping.device === DeviceType.PANASONIC_PTZ) {
 				if (mapping.mappingType === MappingPanasonicPtzType.PRESET) {
-					let tlObjectSource = tlObject as TimelineObjPanasonicPtzPreset
+					let tlObjectSource = tlObject as TimelineResolvedObject & TimelineObjPanasonicPtzPreset
 					_.extend(ptzState, {
 						preset: tlObjectSource.content.preset
 					})
 				} else if (mapping.mappingType === MappingPanasonicPtzType.PRESET_SPEED) {
-					let tlObjectSource = tlObject as TimelineObjPanasonicPtzPresetSpeed
+					let tlObjectSource = tlObject as TimelineResolvedObject & TimelineObjPanasonicPtzPresetSpeed
 					_.extend(ptzState, {
 						speed: tlObjectSource.content.speed
 					})
 				} else if (mapping.mappingType === MappingPanasonicPtzType.ZOOM_SPEED) {
-					let tlObjectSource = tlObject as TimelineObjPanasonicPtzZoomSpeed
+					let tlObjectSource = tlObject as TimelineResolvedObject & TimelineObjPanasonicPtzZoomSpeed
 					_.extend(ptzState, {
 						zoomSpeed: tlObjectSource.content.zoomSpeed
 					})
 				} else if (mapping.mappingType === MappingPanasonicPtzType.ZOOM) {
-					let tlObjectSource = tlObject as TimelineObjPanasonicPtzZoom
+					let tlObjectSource = tlObject as TimelineResolvedObject & TimelineObjPanasonicPtzZoom
 					_.extend(ptzState, {
 						zoom: tlObjectSource.content.zoom
 					})
