+++ src/index.ts
@@ -1,12 +1,9 @@
 
 export * from './conductor'
-export * from './devices/mapping'
+export * from './types/mapping'
 
 // let myConductor = new Conductor();
 export { DeviceOptions } from './devices/device'
 export { CasparCGDevice } from './devices/casparCG'
 
-export * from './enums'
-export * from './types'
-
-// @todo remove this line of comment
+export * from './types/index'
