+++ src/types/hyperdeck.ts
@@ -1,0 +1,14 @@
+import { Mapping, DeviceType } from './mapping'
+
+export interface MappingHyperdeck extends Mapping {
+	device: DeviceType.HYPERDECK,
+	mappingType: MappingHyperdeckType
+	index?: number
+}
+export enum MappingHyperdeckType {
+	TRANSPORT = 'transport'
+}
+
+export enum TimelineContentTypeHyperdeck {
+	TRANSPORT = 'transport'
+}
